Include default replication options
The README is a little misleading, it specifies defaults for `REPLICATION_CONFIG_SYNCPROV` and `REPLICATION_DB_SYNCPROV`. This patch adds those defaults.

I originally wanted to put these in `/assets/defaults/10-openldap`, but `$BASE_DN` is not known when that file is loaded, so they are included in `configure_ldap_server()`

Note: `starttls=critical` is removed from the option string, I think it should not be hardcoded (convention over configuration)